lint: clear remaining Ruff findings (idioms, exception naming, parametrize)
Manual fixes for the rules that autofix couldn't safely resolve:

- N818: rename exceptions to *Error (UnsupportedSchemaVersionError,
  InvalidSidecarError, UnsupportedFormatError)
- UP042: AlbumState now subclasses StrEnum
- RUF046: drop redundant int(round(...)) in duration readers
- E731/N806: lambda runner_fns -> def; module-level _TERMINAL_STATES
- B905: zip(..., strict=) at every call site
- SIM102/105/108/109, B007, F841: collapse nested ifs, contextlib.suppress,
  ternaries, drop unused bindings
- B008 (FastAPI DI): allowed via flake8-bugbear.extend-immutable-calls
- tests: PT006 tuple parametrize, PT018 split asserts, PT011 raises match,
  RUF013 explicit Optional, RUF015 next(), N817 import alias

ruff check src test  -> clean; 318 tests green.

+++ src/harmonist/formats/_vorbis.py
@@ -132,7 +132,7 @@ class VorbisTagger:
         audio = self._open(path)
         if audio is None or not audio.info.length:
             return None
-        return int(round(audio.info.length * 1000))
+        return round(audio.info.length * 1000)
 
     # ---- write ----
